fix: use alpha 3 where it works still
