+++ examples/expo-example/.rnstorybook/preview.tsx
@@ -1,0 +1,43 @@
+import { View, Appearance } from 'react-native';
+import { withBackgrounds } from '@storybook/addon-ondevice-backgrounds';
+import type { Preview } from '@storybook/react';
+
+const preview: Preview = {
+  decorators: [
+    (Story) => (
+      <View style={{ padding: 8, flex: 1 }}>
+        <Story />
+      </View>
+    ),
+    withBackgrounds,
+  ],
+  parameters: {
+    actions: { argTypesRegex: '^on[A-Z].*' },
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/,
+      },
+    },
+    options: {
+      storySort: {
+        method: 'alphabetical',
+        includeNames: true,
+        order: ['ControlExamples', ['ControlExample'], 'InteractionExample', 'DeepControls'],
+      },
+    },
+    hideFullScreenButton: false,
+    noSafeArea: false,
+    my_param: 'anything',
+    backgrounds: {
+      default: Appearance.getColorScheme() === 'dark' ? 'dark' : 'plain',
+      values: [
+        { name: 'plain', value: 'white' },
+        { name: 'dark', value: '#333' },
+        { name: 'app', value: '#eeeeee' },
+      ],
+    },
+  },
+};
+
+export default preview;
