--- examples/expo-example/.storybook-web/preview.tsx
@@ -1,14 +1,0 @@
-/** @type { import('@storybook/react').Preview } */
-const preview = {
-  parameters: {
-    // actions: { argTypesRegex: '^on[A-Z].*' },
-    controls: {
-      matchers: {
-        color: /(background|color)$/i,
-        date: /Date$/i,
-      },
-    },
-  },
-};
-
-export default preview;
