+++ examples/expo-example/.storybook/preview.tsx
@@ -1,41 +1,12 @@
-import { View, Appearance } from 'react-native';
-import { withBackgrounds } from '@storybook/addon-ondevice-backgrounds';
-import type { Preview } from '@storybook/react';
-
-const preview: Preview = {
-  decorators: [
-    (Story) => (
-      <View style={{ padding: 8, flex: 1 }}>
-        <Story />
-      </View>
-    ),
-    withBackgrounds,
-  ],
+/** @type { import('@storybook/react').Preview } */
+const preview = {
   parameters: {
-    actions: { argTypesRegex: '^on[A-Z].*' },
+    // actions: { argTypesRegex: '^on[A-Z].*' },
     controls: {
       matchers: {
         color: /(background|color)$/i,
-        date: /Date$/,
+        date: /Date$/i,
       },
-    },
-    options: {
-      storySort: {
-        method: 'alphabetical',
-        includeNames: true,
-        order: ['ControlExamples', ['ControlExample'], 'InteractionExample', 'DeepControls'],
-      },
-    },
-    hideFullScreenButton: false,
-    noSafeArea: false,
-    my_param: 'anything',
-    backgrounds: {
-      default: Appearance.getColorScheme() === 'dark' ? 'dark' : 'plain',
-      values: [
-        { name: 'plain', value: 'white' },
-        { name: 'dark', value: '#333' },
-        { name: 'app', value: '#eeeeee' },
-      ],
     },
   },
 };
