+++ examples/expo-example/App.tsx
@@ -20,7 +20,7 @@ function App() {
 let AppEntryPoint = App;
 
 if (process.env.EXPO_PUBLIC_STORYBOOK_ENABLED === 'true') {
-  AppEntryPoint = require('./.storybook').default;
+  AppEntryPoint = require('./.rnstorybook').default;
 }
 
 export default AppEntryPoint;
