+++ examples/expo-example/package.json
@@ -8,9 +8,9 @@
     "ios": "EXPO_PUBLIC_STORYBOOK_ENABLED=true expo start --ios",
     "web": "EXPO_PUBLIC_STORYBOOK_ENABLED=true expo start --web",
     "storybook": "EXPO_PUBLIC_STORYBOOK_ENABLED=true expo start -c",
-    "storybook:web": "storybook dev -p 6006 -c ./.storybook-web",
-    "build-web-storybook": "storybook build -c ./.storybook-web",
-    "storybook-generate": "sb-rn-get-stories --config-path=./.storybook",
+    "storybook:web": "storybook dev -p 6006",
+    "build-web-storybook": "storybook build",
+    "storybook-generate": "sb-rn-get-stories --config-path=./.rnstorybook",
     "disabled-example": "expo start",
     "format": "prettier --write .",
     "lint": "eslint . --ext .js,.jsx,.ts,.tsx",
@@ -27,20 +27,20 @@
     "@react-native-async-storage/async-storage": "1.23.1",
     "@react-native-community/datetimepicker": "8.2.0",
     "@react-native-community/slider": "4.5.5",
-    "@storybook/addon-essentials": "9.0.0-alpha.4",
-    "@storybook/addon-interactions": "9.0.0-alpha.4",
-    "@storybook/addon-links": "9.0.0-alpha.4",
+    "@storybook/addon-essentials": "9.0.0-alpha.3",
+    "@storybook/addon-interactions": "9.0.0-alpha.3",
+    "@storybook/addon-links": "9.0.0-alpha.3",
     "@storybook/addon-ondevice-actions": "^8.6.1",
     "@storybook/addon-ondevice-backgrounds": "^8.6.1",
     "@storybook/addon-ondevice-controls": "^8.6.1",
     "@storybook/addon-ondevice-notes": "^8.6.1",
     "@storybook/addon-react-native-server": "0.0.6",
-    "@storybook/blocks": "9.0.0-alpha.4",
+    "@storybook/blocks": "9.0.0-alpha.3",
     "@storybook/global": "^5.0.0",
-    "@storybook/react": "9.0.0-alpha.4",
+    "@storybook/react": "9.0.0-alpha.3",
     "@storybook/react-native": "^8.6.1",
     "@storybook/react-native-theming": "^8.6.1",
-    "@storybook/react-native-web-vite": "9.0.0-alpha.4",
+    "@storybook/react-native-web-vite": "9.0.0-alpha.3",
     "expo": "~52.0.39",
     "history": "^5.3.0",
     "querystring": "^0.2.1",
@@ -53,7 +53,7 @@
     "react-native-svg": "15.8.0",
     "react-native-web": "~0.19.13",
     "react-router": "^6.26.2",
-    "storybook": "9.0.0-alpha.4",
+    "storybook": "9.0.0-alpha.3",
     "storybook-addon-deep-controls": "^0.9.2",
     "ws": "^8.18.0"
   },
