+++ packages/ondevice-actions/src/components/ActionLogger/index.tsx
@@ -1,4 +1,4 @@
-import { ActionDisplay } from 'storybook/actions';
+import { ActionDisplay } from '@storybook/addon-actions';
 import { Button, ScrollView, StyleSheet, Text, View } from 'react-native';
 import Inspect from './Inspect';
