+++ packages/ondevice-actions/src/containers/ActionLogger/index.tsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from 'react';
 import deepEqual from 'fast-deep-equal';
-import { addons } from 'storybook/manager-api';
+import { addons } from 'storybook/internal/manager-api';
 import { SET_CURRENT_STORY } from 'storybook/internal/core-events';
-import { ActionDisplay, EVENT_ID } from 'storybook/actions';
+import { ActionDisplay, EVENT_ID } from '@storybook/addon-actions';
 import { ActionLogger as ActionLoggerComponent } from '../../components/ActionLogger';
 
 interface ActionLoggerProps {
