+++ packages/ondevice-backgrounds/src/register.tsx
@@ -1,4 +1,4 @@
-import { addons, types } from 'storybook/manager-api';
+import { addons, types } from 'storybook/internal/manager-api';
 import BackgroundPanel from './BackgroundPanel';
 import { ADDON_ID, PANEL_ID, PARAM_KEY } from './constants';
