+++ packages/ondevice-controls/package.json
@@ -30,7 +30,7 @@
     "copyimages": "cross-env-shell cp -r src/components/color-picker/resources dist/components/color-picker/resources"
   },
   "dependencies": {
-    "@storybook/addon-controls": "9.0.0-alpha.4",
+    "@storybook/addon-controls": "9.0.0-alpha.3",
     "@storybook/react-native-theming": "^8.6.1",
     "@storybook/react-native-ui": "^8.6.1",
     "deep-equal": "^1.0.1",
