+++ packages/ondevice-controls/src/index.tsx
@@ -1,4 +1,4 @@
-import { addons, types } from 'storybook/manager-api';
+import { addons, types } from 'storybook/internal/manager-api';
 
 import ControlsPanel from './ControlsPanel';
 import { AddonPanel } from './Panel';
