+++ packages/react-native-ui/src/Layout.tsx
@@ -1,5 +1,5 @@
 import { SET_CURRENT_STORY } from 'storybook/internal/core-events';
-import { addons } from 'storybook/manager-api';
+import { addons } from 'storybook/internal/manager-api';
 import { type API_IndexHash, type Args, type StoryContext } from 'storybook/internal/types';
 import type { ReactRenderer } from '@storybook/react';
 import { styled, useTheme } from '@storybook/react-native-theming';
