+++ packages/react-native/package.json
@@ -47,7 +47,7 @@
   "dependencies": {
     "@storybook/csf": "^0.1.13",
     "@storybook/global": "^5.0.0",
-    "@storybook/react": "9.0.0-alpha.4",
+    "@storybook/react": "9.0.0-alpha.3",
     "@storybook/react-native-theming": "^8.6.1",
     "@storybook/react-native-ui": "^8.6.1",
     "commander": "^8.2.0",
@@ -56,7 +56,7 @@
     "prettier": "^2.4.1",
     "react-native-url-polyfill": "^2.0.0",
     "setimmediate": "^1.0.5",
-    "storybook": "9.0.0-alpha.4",
+    "storybook": "9.0.0-alpha.3",
     "type-fest": "~2.19",
     "util": "^0.12.4",
     "ws": "^8.18.0"
